HAAR-4916: ✨ update app to use kotlin library

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/model/PrisonUserDetails.kt
@@ -43,7 +43,7 @@ data class PrisonUserDetails(
     userId = userId.toString(),
     name = name,
     uuid = null,
-    staffId = userId.toLong(),
+    staffId = userId,
     activeCaseLoadId = activeCaseloadId,
   )
